move file structure
